Add 3 reports to dropdown and create format dialog box
This adds the 'download reports' menu to the main event page (it
was already on the edit list page), and refactors it to list the
three available reports. When a menu item is clicked a modal
dialog box opens for the user to choose between CSV and wikitext
formats. Menu items are disabled if the event data is not
available yet (i.e. the event hasn't been updated).

Bug: T206576